Got committee/candidate relationships working
- Had to move models into a common file so that they could reference
  each other without it being an error-prone PITA
- Simplified querying
- Altered dynamic filters to work with the join table condition
