+++ data/sql_updates/create_candidates_vw.sql
@@ -12,7 +12,7 @@ select
         else 'Unknown' end as candidate_status_full,
     max(dimoffice.office_district) as district,
     csi_recent.election_yr as active_through,
-    (select array_agg(distinct election_yr)::int[] from dimcandstatusici csi where csi.cand_sk = dimcand.cand_sk) as election_years,
+    array_agg(distinct csi_all.election_yr)::int[] as election_years,
     max(csi_recent.ici_code) as incumbent_challenge,
     case max(csi_recent.ici_code)
         when 'I' then 'Incumbent'
@@ -26,8 +26,9 @@ select
     max(dimoffice.office_state) as state,
     (select cand_nm from dimcandproperties cp where cp.cand_sk = dimcand.cand_sk order by candproperties_sk desc limit 1) as name
 from dimcand
-    inner join (select distinct on (cand_sk) cand_sk, election_yr, cand_status, ici_code from dimcandstatusici order by cand_sk, election_yr desc) csi_recent using (cand_sk)
-    inner join dimcandoffice co on co.cand_sk = dimcand.cand_sk and co.cand_election_yr = csi_recent.election_yr  -- only joined to get to dimoffice
+    left join (select distinct on (cand_sk) cand_sk, election_yr, cand_status, ici_code from dimcandstatusici order by cand_sk, election_yr desc) csi_recent using (cand_sk)
+    left join dimcandstatusici csi_all using (cand_sk)
+    inner join dimcandoffice co on co.cand_sk = dimcand.cand_sk and (csi_recent.election_yr is null or co.cand_election_yr = csi_recent.election_yr)  -- only joined to get to dimoffice
     inner join dimoffice using (office_sk)
     inner join dimparty using (party_sk)
 group by
@@ -35,4 +36,3 @@ group by
     dimcand.cand_id,
     csi_recent.election_yr
 ;
-grant select on table ofec_candidates_vw to webro;
