+++ webservices/resources/candidates.py
@@ -1,7 +1,6 @@
 from flask.ext.restful import Resource, reqparse, fields, marshal_with, inputs
-from webservices.common.models import db
+from webservices.common.models import db, Candidate
 from webservices.common.util import default_year
-from sqlalchemy.dialects.postgresql import ARRAY
 from sqlalchemy.sql import text
 
 
@@ -110,27 +109,3 @@ class CandidateList(Resource):
         return count, candidates.order_by(Candidate.name).paginate(page_num, per_page, False).items
 
 
-class Candidate(db.Model):
-    candidate_key = db.Column(db.Integer, primary_key=True)
-    candidate_id = db.Column(db.String(10))
-    candidate_status = db.Column(db.String(1))
-    candidate_status_full = db.Column(db.String(11))
-    district = db.Column(db.String(2))
-    active_through = db.Column(db.Integer)
-    election_years = db.Column(ARRAY(db.Integer))
-    incumbent_challenge = db.Column(db.String(1))
-    incumbent_challenge_full = db.Column(db.String(10))
-    office = db.Column(db.String(1))
-    office_full = db.Column(db.String(9))
-    party = db.Column(db.String(3))
-    party_full = db.Column(db.String(255))
-    state = db.Column(db.String(2))
-    name = db.Column(db.String(100))
-
-    __tablename__ = 'ofec_candidates_vw'
-
-class CandidateFulltext(db.Model):
-    cand_sk = db.Column(db.Integer, primary_key=True)
-    fulltxt = db.Column(db.Text)
-
-    __tablename__ = 'dimcand_fulltext'
