+++ webservices/rest.py
@@ -132,7 +132,7 @@ api.add_resource(Help, '/')
 api.add_resource(CandidateResource, '/candidate/<string:id>')
 api.add_resource(CandidateList, '/candidate')
 api.add_resource(CommitteeResource, '/committee/<string:id>')
-api.add_resource(CommitteeList, '/committee')
+api.add_resource(CommitteeList, '/committee', '/candidate/<string:id>/committees')
 api.add_resource(TotalResource, '/total/<string:id>')
 api.add_resource(TotalSearch, '/total')
 api.add_resource(NameSearch, '/name')
